% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_twolayer.R
\name{nn_twolayer}
\alias{nn_twolayer}
\title{Neural Network with Two Layers}
\usage{
nn_twolayer(Words = 3000, Text = man_dat2$text,
  Codes = man_dat2$cmp_code2, Epochs = 2, Weighting = "count",
  Seed = 17, Batch = 32)
}
\arguments{
\item{Words}{The number of top words included in document feature matrixes used as training and testing data.}

\item{Text}{The text that will be used as training and test data.}

\item{Codes}{The codes that will be used as outcomes to be predicted by the NN model.}

\item{Epochs}{The number of epochs used in the NN model.}

\item{Weighting}{The type of feature weighting used in the document feature matrix. I.e., count and tfidf.}

\item{Seed}{The seed used in the model. Defaults to 17}

\item{Batch}{The number of batches estimated in the NN.}
}
\description{
This function is a wrapper for a two layered neural network written using the Keras Package. It takes a
}
\keyword{networks}
\keyword{neural}
