% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{packages}
\alias{packages}
\title{Packages: a function for loading packages}
\usage{
packages(lib = NA)
}
\arguments{
\item{lib}{A string character containing the names of CRAN packages to be loaded or installed in R}
}
\description{
A function designed to simplify loading and installing functions
}
