% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_embed_lstm_function.R
\name{nn_embedded_lstm}
\alias{nn_embedded_lstm}
\title{Neural Network with a long-term short-term memory and an embedding layer}
\usage{
nn_embedded_lstm(Words = 3000, Text = man_dat2$text,
  Codes = man_dat2$cmp_code2, Epochs = 10,
  Seed = as.numeric(Sys.time()), Batch = 32, MaxSentencelen = 60,
  WordEmbedDim = 50, ValSplit = 0.1, CM = TRUE, Model = FALSE)
}
\arguments{
\item{Words}{The number of top words included in document feature matrixes used as training and testing data.}

\item{Text}{The text that will be used as training and test data.}

\item{Codes}{The codes that will be used as outcomes to be predicted by the NN model.}

\item{Epochs}{The number of epochs used in the NN model.}

\item{Seed}{The seed used in the model. Defaults to 17}

\item{Batch}{The number of batches estimated in the NN.}

\item{MaxSentencelen}{All sentences will be truncated to this length to be input into the LSTM model}

\item{WordEmbedDim}{The number of word embedding dimensions to be produced by the LSTM model}

\item{ValSplit}{The validation split of the data used in the training of the LSTM model}

\item{CM}{A logical variable that indicates whether a confusion matrix will be output from the function}

\item{Model}{A logical variable that indicates whether the trained model should be included in the output of this function}
}
\description{
This function is a wrapper for a long term short term neural network written using the Keras Package.
}
\keyword{LSTM}
\keyword{networks,}
\keyword{neural}
